new card design added to hpc command
