+++ app/utilities/drawing/badges/daru69.png
