+++ app/utilities/drawing/badges/operation_elysian_vet.png
