+++ app/utilities/drawing/fonts/MerriweatherSansItalic/OFL.txt
