+++ app/utilities/drawing/fonts/MerriweatherSansItalic/static/MerriweatherSans-Bold.ttf
