+++ app/utilities/drawing/fonts/MerriweatherSansItalic/static/MerriweatherSans-BoldItalic.ttf
