+++ app/utilities/drawing/fonts/MerriweatherSansItalic/static/MerriweatherSans-ExtraBold.ttf
