+++ app/utilities/drawing/fonts/MerriweatherSansItalic/static/MerriweatherSans-ExtraBoldItalic.ttf
