+++ app/utilities/drawing/fonts/MerriweatherSansItalic/static/MerriweatherSans-Italic.ttf
