+++ app/utilities/drawing/fonts/MerriweatherSansItalic/static/MerriweatherSans-Light.ttf
