+++ app/utilities/drawing/fonts/MerriweatherSansItalic/static/MerriweatherSans-LightItalic.ttf
