+++ app/utilities/drawing/fonts/MerriweatherSansItalic/static/MerriweatherSans-Medium.ttf
