+++ app/utilities/drawing/fonts/MerriweatherSansItalic/static/MerriweatherSans-MediumItalic.ttf
