+++ app/utilities/drawing/fonts/MerriweatherSansItalic/static/MerriweatherSans-Regular.ttf
