+++ app/utilities/drawing/fonts/MerriweatherSansItalic/static/MerriweatherSans-SemiBold.ttf
