+++ app/utilities/drawing/fonts/MerriweatherSansItalic/static/MerriweatherSans-SemiBoldItalic.ttf
