--- app/utilities/drawingPlayground/__init__.py
@@ -1,1 +1,0 @@
-
