--- app/utilities/drawingPlayground/constants.py
@@ -1,35 +1,0 @@
-# CONSTANTS THAT ARE USED FOR drawing/
-# THEY DEFINE PLACEMENTS/COLORS AND DEFAULT FONTS/IMAGES PATHS
-
-
-# SIZES
-background_dimensions = (419, 404)
-blurry_region_radius = 12
-blurry_region_size = (395, 380)
-black_region_1_size = (371, 118)
-black_region_2_size = (371, 88)
-badge_size = (41,41)
-pfp_circle_size = (100,100)
-xp_bar_size = (283, 10)
-
-# OFFSETS
-blurry_background_offset = (12,12)
-black_region_1_offset = (24, 164)
-black_region_2_offset = (24, 292)
-first_badge_offset = (36, 329)
-badges_spacing = 10
-pfp_circle_offset = (295,39)
-xp_bar_offset = (100, 195)
-
-# COLORS
-
-
-# IMAGE PATHS
-server_icon_path = 'server_icon.png'
-badges_map = {
-    "operation_elysian_veteran": "badges/operation_elysian_vet.png",
-    "daru69": "badges/daru69.png"
-}
-
-# FONTS
-default_font = "fonts/MerriweatherSansItalic/MerriweatherSans-VariableFont_wght.ttf"
